+++ internal/core/graphics/wgpu_26.rs
@@ -10,7 +10,7 @@ in particular the `BackendSelector` type, to configure the WGPU-based renderer(s
 
 use alloc::boxed::Box;
 
-pub use wgpu_25 as wgpu;
+pub use wgpu_26 as wgpu;
 
 pub mod api {
     /*!
@@ -24,49 +24,52 @@ pub mod api {
     #[non_exhaustive]
     pub struct WGPUSettings {
         /// The backends to use for the WGPU instance.
-        pub backends: wgpu_25::Backends,
+        pub backends: wgpu_26::Backends,
         /// The different options that are given to the selected backends.
-        pub backend_options: wgpu_25::BackendOptions,
+        pub backend_options: wgpu_26::BackendOptions,
         /// The flags to fine-tune behaviour of the WGPU instance.
-        pub instance_flags: wgpu_25::InstanceFlags,
+        pub instance_flags: wgpu_26::InstanceFlags,
+        /// Memory budget thresholds used by some backends.
+        pub instance_memory_budget_thresholds: wgpu_26::MemoryBudgetThresholds,
 
         /// The power preference is used to influence the WGPU adapter selection.
-        pub power_preference: wgpu_25::PowerPreference,
+        pub power_preference: wgpu_26::PowerPreference,
 
         /// The label for the device. This is used to identify the device in debugging tools.
         pub device_label: Option<std::borrow::Cow<'static, str>>,
         /// The required features for the device.
-        pub device_required_features: wgpu_25::Features,
+        pub device_required_features: wgpu_26::Features,
         /// The required limits for the device.
-        pub device_required_limits: wgpu_25::Limits,
+        pub device_required_limits: wgpu_26::Limits,
         /// The memory hints for the device.
-        pub device_memory_hints: wgpu_25::MemoryHints,
+        pub device_memory_hints: wgpu_26::MemoryHints,
     }
 
     impl Default for WGPUSettings {
         fn default() -> Self {
-            let backends = wgpu_25::Backends::from_env().unwrap_or_default();
-            let dx12_shader_compiler = wgpu_25::Dx12Compiler::from_env().unwrap_or_default();
-            let gles_minor_version = wgpu_25::Gles3MinorVersion::from_env().unwrap_or_default();
+            let backends = wgpu_26::Backends::from_env().unwrap_or_default();
+            let dx12_shader_compiler = wgpu_26::Dx12Compiler::from_env().unwrap_or_default();
+            let gles_minor_version = wgpu_26::Gles3MinorVersion::from_env().unwrap_or_default();
 
             Self {
                 backends,
-                backend_options: wgpu_25::BackendOptions {
-                    dx12: wgpu_25::Dx12BackendOptions { shader_compiler: dx12_shader_compiler },
-                    gl: wgpu_25::GlBackendOptions {
+                backend_options: wgpu_26::BackendOptions {
+                    dx12: wgpu_26::Dx12BackendOptions { shader_compiler: dx12_shader_compiler },
+                    gl: wgpu_26::GlBackendOptions {
                         gles_minor_version,
-                        fence_behavior: wgpu_25::GlFenceBehavior::default(),
+                        fence_behavior: wgpu_26::GlFenceBehavior::default(),
                     },
                     noop: wgpu::NoopBackendOptions::default(),
                 },
-                instance_flags: wgpu_25::InstanceFlags::from_build_config().with_env(),
+                instance_flags: wgpu_26::InstanceFlags::from_build_config().with_env(),
+                instance_memory_budget_thresholds: wgpu_26::MemoryBudgetThresholds::default(),
 
-                power_preference: wgpu_25::PowerPreference::from_env().unwrap_or_default(),
+                power_preference: wgpu_26::PowerPreference::from_env().unwrap_or_default(),
 
                 device_label: None,
-                device_required_features: wgpu_25::Features::empty(),
-                device_required_limits: wgpu_25::Limits::downlevel_webgl2_defaults(),
-                device_memory_hints: wgpu_25::MemoryHints::MemoryUsage,
+                device_required_features: wgpu_26::Features::empty(),
+                device_required_limits: wgpu_26::Limits::downlevel_webgl2_defaults(),
+                device_memory_hints: wgpu_26::MemoryHints::MemoryUsage,
             }
         }
     }
@@ -79,13 +82,13 @@ pub mod api {
         /// device, and queue for use.
         Manual {
             /// The WGPU instance to use.
-            instance: wgpu_25::Instance,
+            instance: wgpu_26::Instance,
             /// The WGPU adapter to use.
-            adapter: wgpu_25::Adapter,
+            adapter: wgpu_26::Adapter,
             /// The WGPU device to use.
-            device: wgpu_25::Device,
+            device: wgpu_26::Device,
             /// The WGPU queue to use.
-            queue: wgpu_25::Queue,
+            queue: wgpu_26::Queue,
         },
         /// Use `Automatic` if you want to let Slint select the WGPU instance, adapter, and
         /// device, but fine-tune aspects such as memory limits or features.
@@ -98,23 +101,23 @@ pub mod api {
         }
     }
 
-    impl TryFrom<wgpu_25::Texture> for super::super::Image {
+    impl TryFrom<wgpu_26::Texture> for super::super::Image {
         type Error = TextureImportError;
 
-        fn try_from(texture: wgpu_25::Texture) -> Result<Self, Self::Error> {
-            if texture.format() != wgpu_25::TextureFormat::Rgba8Unorm
-                && texture.format() != wgpu_25::TextureFormat::Rgba8UnormSrgb
+        fn try_from(texture: wgpu_26::Texture) -> Result<Self, Self::Error> {
+            if texture.format() != wgpu_26::TextureFormat::Rgba8Unorm
+                && texture.format() != wgpu_26::TextureFormat::Rgba8UnormSrgb
             {
                 return Err(Self::Error::InvalidFormat);
             }
             let usages = texture.usage();
-            if !usages.contains(wgpu_25::TextureUsages::TEXTURE_BINDING)
-                || !usages.contains(wgpu_25::TextureUsages::RENDER_ATTACHMENT)
+            if !usages.contains(wgpu_26::TextureUsages::TEXTURE_BINDING)
+                || !usages.contains(wgpu_26::TextureUsages::RENDER_ATTACHMENT)
             {
                 return Err(Self::Error::InvalidUsage);
             }
             Ok(Self(super::super::ImageInner::WGPUTexture(
-                super::super::WGPUTexture::WGPU25Texture(texture),
+                super::super::WGPUTexture::WGPU26Texture(texture),
             )))
         }
     }
@@ -153,16 +156,16 @@ pub fn init_instance_adapter_device_queue_surface(
     requested_graphics_api: Option<RequestedGraphicsAPI>,
 ) -> Result<
     (
-        wgpu_25::Instance,
-        wgpu_25::Adapter,
-        wgpu_25::Device,
-        wgpu_25::Queue,
-        wgpu_25::Surface<'static>,
+        wgpu_26::Instance,
+        wgpu_26::Adapter,
+        wgpu_26::Device,
+        wgpu_26::Queue,
+        wgpu_26::Surface<'static>,
     ),
     Box<dyn std::error::Error + Send + Sync + 'static>,
 > {
     let (instance, adapter, device, queue, surface) = match requested_graphics_api {
-        Some(RequestedGraphicsAPI::WGPU25(api::WGPUConfiguration::Manual {
+        Some(RequestedGraphicsAPI::WGPU26(api::WGPUConfiguration::Manual {
             instance,
             adapter,
             device,
@@ -171,14 +174,15 @@ pub fn init_instance_adapter_device_queue_surface(
             let surface = instance.create_surface(window_handle).unwrap();
             (instance, adapter, device, queue, surface)
         }
-        Some(RequestedGraphicsAPI::WGPU25(api::WGPUConfiguration::Automatic(wgpu25_settings))) => {
+        Some(RequestedGraphicsAPI::WGPU26(api::WGPUConfiguration::Automatic(wgpu26_settings))) => {
             // wgpu uses async here, but the returned future is ready on first poll on all platforms except WASM,
             // which we don't support right now.
             let instance = poll_once(async {
                 wgpu::util::new_instance_with_webgpu_detection(&wgpu::InstanceDescriptor {
-                    backends: wgpu25_settings.backends,
-                    flags: wgpu25_settings.instance_flags,
-                    backend_options: wgpu25_settings.backend_options,
+                    backends: wgpu26_settings.backends,
+                    flags: wgpu26_settings.instance_flags,
+                    backend_options: wgpu26_settings.backend_options,
+                    memory_budget_thresholds: wgpu26_settings.instance_memory_budget_thresholds,
                 })
                 .await
             })
@@ -194,7 +198,7 @@ pub fn init_instance_adapter_device_queue_surface(
                     Err(_) => {
                         instance
                             .request_adapter(&wgpu::RequestAdapterOptions {
-                                power_preference: wgpu25_settings.power_preference,
+                                power_preference: wgpu26_settings.power_preference,
                                 force_fallback_adapter: false,
                                 compatible_surface: Some(&surface),
                             })
@@ -208,13 +212,13 @@ pub fn init_instance_adapter_device_queue_surface(
             let (device, queue) = poll_once(async {
                 adapter
                     .request_device(&wgpu::DeviceDescriptor {
-                        label: wgpu25_settings.device_label.as_deref(),
-                        required_features: wgpu25_settings.device_required_features,
+                        label: wgpu26_settings.device_label.as_deref(),
+                        required_features: wgpu26_settings.device_required_features,
                         // Make sure we use the texture resolution limits from the adapter, so we can support images the size of the swapchain.
-                        required_limits: wgpu25_settings
+                        required_limits: wgpu26_settings
                             .device_required_limits
                             .using_resolution(adapter.limits()),
-                        memory_hints: wgpu25_settings.device_memory_hints,
+                        memory_hints: wgpu26_settings.device_memory_hints,
                         trace: wgpu::Trace::default(),
                     })
                     .await
@@ -243,6 +247,7 @@ pub fn init_instance_adapter_device_queue_surface(
                         },
                         noop: wgpu::NoopBackendOptions::default(),
                     },
+                    memory_budget_thresholds: wgpu::MemoryBudgetThresholds::default(),
                 })
                 .await
             })